+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 - `org.Party`: avoid panic when regime's normalizer is not present
 - `br`: set missing normalizer in regime definition
+- `tax`: `Normalizers()` method in `RegimeDef`
 
 ## [v0.220.5] - 2025-07-21